backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
SQLAlchemy>=2
alembic
email-validator
python-multipart
pyodbc
greenlet
passlib[bcrypt]
python-jose[cryptography]
python-slugify
httpx
pytest
pytest-asyncio
ruff
black
isort
mypy
